refactor: isolate anthropic stream projection
