+++ scripts/rust-source-size-baseline.txt
@@ -1,6 +1,5 @@
 # Rust source files above the 1,000-line target. Keep sorted by path.
 # A file may only stay equal to this maximum or shrink with this entry tightened.
-crates/llm/src/anthropic_messages.rs 1775
 crates/llm/src/chatgpt_responses.rs 1712
 crates/llm/src/google_gemini_generate_content.rs 1373
 crates/llm/src/lib.rs 1579
